+++ typechain/factories/contracts/libraries/PositionManager__factory.ts
@@ -260,7 +260,7 @@ const _abi = [
               {
                 internalType: "Epoch",
                 name: "epoch",
-                type: "uint8",
+                type: "uint16",
               },
             ],
             internalType: "struct CouponKey",
@@ -451,7 +451,7 @@ const _abi = [
               {
                 internalType: "Epoch",
                 name: "epoch",
-                type: "uint8",
+                type: "uint16",
               },
             ],
             internalType: "struct CouponKey",
